pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "zotero-client"
version = "0.1.1"
authors = [
  { name="Your Name", email="your.email@example.com" },
]
description = "A Python client for interacting with the Zotero API"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "openai>=1.0.0",
    "numpy>=1.24.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/zotero-app"

[project.scripts]
cl = "zotero_client.cli.main:main"

[tool.setuptools.packages.find]
include = ["cli*", "zotero*"]
